requirements.txt
numpy>=1.9.2
scikit-learn>=0.22.0
matplotlib>=1.4.3
pandas>=0.19
streamlit>=0.74.1
keras >= 2.0
tensorflow-cpu >=2.5.0
